Lfs (#12)
* refactor list files

* check if the file is lfs
