+++ modules/structs/repo_file.go
@@ -4,8 +4,6 @@
 
 package structs
 
-import "time"
-
 // FileOptions options for all file APIs
 type FileOptions struct {
 	// message (optional) for the commit of this file. if not supplied, a default message will be used
@@ -123,32 +121,6 @@ type ContentsResponse struct {
 	Path          string `json:"path"`
 	SHA           string `json:"sha"`
 	LastCommitSHA string `json:"last_commit_sha"`
-	// `type` will be `file`, `dir`, `symlink`, or `submodule`
-	Type string `json:"type"`
-	Size int64  `json:"size"`
-	// `encoding` is populated when `type` is `file`, otherwise null
-	Encoding *string `json:"encoding"`
-	// `content` is populated when `type` is `file`, otherwise null
-	Content *string `json:"content"`
-	// `target` is populated when `type` is `symlink`, otherwise null
-	Target      *string `json:"target"`
-	URL         *string `json:"url"`
-	HTMLURL     *string `json:"html_url"`
-	GitURL      *string `json:"git_url"`
-	DownloadURL *string `json:"download_url"`
-	// `submodule_git_url` is populated when `type` is `submodule`, otherwise null
-	SubmoduleGitURL *string            `json:"submodule_git_url"`
-	Links           *FileLinksResponse `json:"_links"`
-}
-
-// CommitContentsResponse contains information about a repo's entry's (dir, file, symlink, submodule) metadata and content and commit
-type CommitContentsResponse struct {
-	Name              string    `json:"name"`
-	Path              string    `json:"path"`
-	SHA               string    `json:"sha"`
-	LastCommitSHA     string    `json:"last_commit_sha"`
-	LastCommitMessage string    `json:"last_commit_message"`
-	LastCommitCreate  time.Time `json:"last_commit_create"`
 	// `type` will be `file`, `dir`, `symlink`, or `submodule`
 	Type string `json:"type"`
 	Size int64  `json:"size"`
